Refactor driver packages
