+++ oshi-core/src/main/java/oshi/driver/unix/solaris/Who.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package oshi.driver.mac.net;
+package oshi.driver.unix.solaris;
 
 import static oshi.jna.platform.unix.CLibrary.LOGIN_PROCESS;
 import static oshi.jna.platform.unix.CLibrary.USER_PROCESS;
@@ -31,8 +31,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 import oshi.annotation.concurrent.ThreadSafe;
-import oshi.jna.platform.mac.SystemB;
-import oshi.jna.platform.mac.SystemB.MacUtmpx;
+import oshi.jna.platform.unix.solaris.SolarisLibc;
+import oshi.jna.platform.unix.solaris.SolarisLibc.SolarisUtmpx;
 import oshi.software.os.OSSession;
 
 /**
@@ -41,7 +41,7 @@ import oshi.software.os.OSSession;
 @ThreadSafe
 public final class Who {
 
-    private static final SystemB SYS = SystemB.INSTANCE;
+    private static final SolarisLibc LIBC = SolarisLibc.INSTANCE;
 
     private Who() {
     }
@@ -53,21 +53,23 @@ public final class Who {
      */
     public static synchronized List<OSSession> queryUtxent() {
         List<OSSession> whoList = new ArrayList<>();
-        MacUtmpx ut;
+        SolarisUtmpx ut;
         // Rewind
-        SYS.setutxent();
+        LIBC.setutxent();
         // Iterate
-        while ((ut = SYS.getutxent()) != null) {
+        while ((ut = LIBC.getutxent()) != null) {
             if (ut.ut_type == USER_PROCESS || ut.ut_type == LOGIN_PROCESS) {
                 String user = new String(ut.ut_user, StandardCharsets.US_ASCII).trim();
-                String device = new String(ut.ut_line, StandardCharsets.US_ASCII).trim();
-                String host = new String(ut.ut_host, StandardCharsets.US_ASCII).trim();
-                long loginTime = ut.ut_tv.tv_sec.longValue() * 1000L + ut.ut_tv.tv_usec / 1000L;
-                whoList.add(new OSSession(user, device, loginTime, host));
+                if (!"LOGIN".equals(user)) {
+                    String device = new String(ut.ut_line, StandardCharsets.US_ASCII).trim();
+                    String host = new String(ut.ut_host, StandardCharsets.US_ASCII).trim();
+                    long loginTime = ut.ut_tv.tv_sec.longValue() * 1000L + ut.ut_tv.tv_usec.longValue() / 1000L;
+                    whoList.add(new OSSession(user, device, loginTime, host));
+                }
             }
         }
         // Close
-        SYS.endutxent();
+        LIBC.endutxent();
         return whoList;
     }
 }
